Finished refactoring!
Now that the code is easier to read and actually focuses on what it
does in a readable way, it is easier to reason about. In this case, we
have blocks being passed around in order to evaluate a bunch of
paths (read: strings) to determine if they're correct. Why do multiple
`start_with?` and `end_with?` calls? It doesn't cleanly describe what
it is trying to do. Instead, we can use regexps and it'll work just
fine, so lets do that. This simplifies things even further.

46.5: flog total

Signed-off-by: Ryan Davis <zenspider@chef.io>